JPERF-1409 Move JfrFilter to api
